requirements.txt
django==1.10
djangorestframework==3.5
Pillow==3.3.1
django-filter
django-crispyforms
